+++ src/NoMatch.tsx
@@ -1,0 +1,5 @@
+const NoMatch = () => {
+  return (<h2>このページは存在しません。</h2>);
+}
+
+export default NoMatch;
